+++ tests/test_orchestrator.py
@@ -18,9 +18,11 @@ from trans_novel.glossary.store import GlossaryStore
 from trans_novel.ingest.models import Chapter, Segment
 from trans_novel.llm.providers.fake import FakeClient
 from trans_novel.llm.usage import UsageSample
+from trans_novel.pipeline.annotations import AnnotationService
 from trans_novel.pipeline.context import RollingContext
-from trans_novel.pipeline.orchestrator import Orchestrator, _BatchResult, _normalize_lang
+from trans_novel.pipeline.orchestrator import Orchestrator
 from trans_novel.pipeline.runstore import STATUS_DONE, STATUS_PENDING, RunStore
+from trans_novel.pipeline.translation import TranslationService, _BatchResult
 
 
 def _translated_para_count(calls) -> int:
@@ -182,7 +184,7 @@ class TestOrchestrator(unittest.TestCase):
             },
         }
 
-        contexts = Orchestrator._annotation_contexts_for_segments(segments, registry)
+        contexts = AnnotationService.annotation_contexts_for_segments(segments, registry)
 
         self.assertEqual(
             [item["target_key"] for item in contexts[0]],
@@ -234,7 +236,7 @@ class TestOrchestrator(unittest.TestCase):
             },
         }
 
-        contexts = Orchestrator._annotation_contexts_for_segments(segments, registry)
+        contexts = AnnotationService.annotation_contexts_for_segments(segments, registry)
 
         self.assertEqual(
             [[item["target_key"] for item in piece] for piece in contexts],
@@ -242,7 +244,7 @@ class TestOrchestrator(unittest.TestCase):
         )
         metadata["source_length"] = 5
         self.assertEqual(
-            Orchestrator._annotation_contexts_for_segments(segments, registry),
+            AnnotationService.annotation_contexts_for_segments(segments, registry),
             [[], []],
         )
 
@@ -317,10 +319,10 @@ class TestOrchestrator(unittest.TestCase):
 
             try:
                 with (
-                    patch.object(orch, "_process_batch", side_effect=process),
+                    patch.object(orch._translation, "process_batch", side_effect=process),
                     patch.object(
-                        orch,
-                        "_extract_batch_glossary",
+                        orch._translation,
+                        "extract_batch_glossary",
                         return_value={
                             "inserted": 0,
                             "conflict": 0,
@@ -329,7 +331,7 @@ class TestOrchestrator(unittest.TestCase):
                         },
                     ),
                     patch.object(
-                        orch.extractor,
+                        orch._runtime.extractor,
                         "extract_and_store",
                         return_value={
                             "inserted": 0,
@@ -339,7 +341,7 @@ class TestOrchestrator(unittest.TestCase):
                         },
                     ),
                 ):
-                    orch._translate_chapter(
+                    orch._translation.translate_chapter(
                         0,
                         store,
                         glossary,
@@ -416,7 +418,7 @@ class TestOrchestrator(unittest.TestCase):
             client = FakeClient(handler=handler)
             orch = Orchestrator(cfg, client=client)
 
-            orch._align_annotations_after_batch(
+            orch._annotations.align_annotations_after_batch(
                 0,
                 chapter,
                 0,
@@ -491,11 +493,11 @@ class TestOrchestrator(unittest.TestCase):
             store = RunStore(os.path.join(directory, "state", "book"))
             orch = Orchestrator(cfg, client=FakeClient(handler=handler))
 
-            orch._align_annotations_after_batch(0, chapter, 0, 1, store)
+            orch._annotations.align_annotations_after_batch(0, chapter, 0, 1, store)
             self.assertEqual(requested, [])
 
             chapter.segments[1].target = "乙"
-            orch._align_annotations_after_batch(0, chapter, 1, 1, store)
+            orch._annotations.align_annotations_after_batch(0, chapter, 1, 1, store)
 
             self.assertEqual(len(requested), 1)
             self.assertIn('"immutable_target": "甲乙"', requested[0])
@@ -573,7 +575,7 @@ class TestOrchestrator(unittest.TestCase):
             store = RunStore(os.path.join(directory, "state", "book"))
             orch = Orchestrator(cfg, client=FakeClient(handler=handler))
 
-            orch._align_annotations_after_batch(0, chapter, 0, 2, store)
+            orch._annotations.align_annotations_after_batch(0, chapter, 0, 2, store)
 
             self.assertEqual(requested_units, ["ch0:a", "ch0:b"])
             saved = store.load_chapter(0)
@@ -660,7 +662,9 @@ class TestOrchestrator(unittest.TestCase):
             client2 = FakeClient(handler=routing_handler)
             orch2 = Orchestrator(cfg, client=client2)
             chapter_indices = [chapter["index"] for chapter in m["chapters"]]
-            expected_total, expected_done = orch2._progress_counts(store, chapter_indices)
+            expected_total, expected_done = orch2._translation.progress_counts(
+                store, chapter_indices
+            )
             progress_events: list[tuple[int, int, str]] = []
             store2 = orch2.run(
                 txt,
@@ -668,7 +672,9 @@ class TestOrchestrator(unittest.TestCase):
             )
             m2 = store2.load_manifest()
             self.assertTrue(all(c["status"] == STATUS_DONE for c in m2["chapters"]))
-            chapter_label = Orchestrator._chapter_progress_label(store.load_chapter(1).title, 1)
+            chapter_label = TranslationService.chapter_progress_label(
+                store.load_chapter(1).title, 1
+            )
             first_chapter_progress = next(
                 event for event in progress_events if event[2] == chapter_label
             )
@@ -800,8 +806,8 @@ class TestSegmentLevelResume(unittest.TestCase):
             snapshot_calls = {"n": 0}
             extract_batch_calls = {"n": 0}
             orch = Orchestrator(cfg, client=FakeClient(handler=self._tr_handler("R2")))
-            real_snapshot = orch._chapter_term_snapshot
-            real_extract = orch._extract_batch_glossary
+            real_snapshot = orch._translation.chapter_term_snapshot
+            real_extract = orch._translation.extract_batch_glossary
 
             def counting_snapshot(glossary, text_segs):
                 snapshot_calls["n"] += 1
@@ -812,8 +818,16 @@ class TestSegmentLevelResume(unittest.TestCase):
                 return real_extract(*args, **kwargs)
 
             with (
-                patch.object(orch, "_chapter_term_snapshot", side_effect=counting_snapshot),
-                patch.object(orch, "_extract_batch_glossary", side_effect=counting_extract),
+                patch.object(
+                    orch._translation,
+                    "chapter_term_snapshot",
+                    side_effect=counting_snapshot,
+                ),
+                patch.object(
+                    orch._translation,
+                    "extract_batch_glossary",
+                    side_effect=counting_extract,
+                ),
             ):
                 orch.run(txt, only_chapter=0)
 
@@ -1311,7 +1325,7 @@ class TestReviewReporting(unittest.TestCase):
             orch = Orchestrator(cfg, client=client)
 
             with (
-                patch("trans_novel.pipeline.orchestrator.load_document") as loader,
+                patch("trans_novel.pipeline.preparation.load_document") as loader,
                 self.assertRaisesRegex(ValueError, "尚无翻译进度"),
             ):
                 orch.run_review(pdf)
@@ -1502,9 +1516,9 @@ class TestReviewReporting(unittest.TestCase):
                 }
 
             with (
-                patch.object(orch, "_review_chapter", side_effect=fake_review),
+                patch.object(orch._review, "review_chapter", side_effect=fake_review),
                 patch(
-                    "trans_novel.pipeline.orchestrator.ReviewConflictArbiter.arbitrate",
+                    "trans_novel.pipeline.review_workflow.ReviewConflictArbiter.arbitrate",
                     new=fake_arbitrate,
                 ),
             ):
@@ -2247,7 +2261,7 @@ class TestReviewReporting(unittest.TestCase):
                     }
                 ]
 
-            with patch.object(orch, "_review_chapter", side_effect=fake_review):
+            with patch.object(orch._review, "review_chapter", side_effect=fake_review):
                 result = orch.run_review(txt)
 
             review_dir = Path(result["review_dir"])
@@ -2330,44 +2344,6 @@ class TestReviewReporting(unittest.TestCase):
 
 
 class TestStyleAnalysis(unittest.TestCase):
-    def _long_doc(self, d):
-        from trans_novel.ingest.segmenter import load_document
-
-        txt = os.path.join(d, "long.txt")
-        chapters = []
-        for i in range(3):
-            # 段落勿以「第N章」开头，避免被 TXT reader 的章标题启发式误判
-            body = "\n\n".join(f"章{i}の段落{j}です。" + "あ" * 60 for j in range(8))
-            chapters.append(f"# 第{i}章\n\n{body}")
-        with open(txt, "w", encoding="utf-8") as f:
-            f.write("\n\n".join(chapters))
-        return load_document(txt, "ja", "zh")
-
-    def test_sample_text_multipoint(self):
-        """labeled=True 多点采样带三个标注；labeled=False 为纯源文单段。"""
-        with tempfile.TemporaryDirectory() as d:
-            doc = self._long_doc(d)
-            labeled = Orchestrator._sample_text(doc)
-            for tag in ("【开头样章】", "【中部样章】", "【结尾样章】"):
-                self.assertIn(tag, labeled)
-            plain = Orchestrator._sample_text(doc, labeled=False)
-            self.assertNotIn("样章】", plain)
-            self.assertIn("章0の段落0です", plain)
-
-    def test_sample_text_short_book_dedup(self):
-        """单章书：三个采样点重合，只取一次、不重复。"""
-        with tempfile.TemporaryDirectory() as d:
-            from trans_novel.ingest.segmenter import load_document
-
-            txt = os.path.join(d, "short.txt")
-            with open(txt, "w", encoding="utf-8") as f:
-                f.write("# 唯一章\n\n" + "长段落。" + "あ" * 300)
-            doc = load_document(txt, "ja", "zh")
-            sample = Orchestrator._sample_text(doc)
-            self.assertEqual(sample.count("【开头样章】"), 1)
-            self.assertNotIn("【中部样章】", sample)
-            self.assertNotIn("【结尾样章】", sample)
-
     def test_style_brief_new_fields(self):
         """style_brief 渲染新风格维度；旧 analysis（缺新字段）不报错不输出。"""
         from trans_novel.agents.analyzer import Analyzer
@@ -2638,22 +2614,11 @@ class TestTierRouting(unittest.TestCase):
             self.assertEqual(seen, set(expect), "各类调用都应出现")
 
 
-class TestLangNormalize(unittest.TestCase):
-    def test_normalize_lang(self):
-        self.assertEqual(_normalize_lang("Japanese"), "ja")
-        self.assertEqual(_normalize_lang("日语"), "ja")
-        self.assertEqual(_normalize_lang("RU"), "ru")
-        self.assertEqual(_normalize_lang("russian"), "ru")
-        self.assertEqual(_normalize_lang("fr"), "fr")
-        self.assertEqual(_normalize_lang("unknown"), "")
-        self.assertEqual(_normalize_lang(""), "")
-
-
 class TestProgressLabels(unittest.TestCase):
     def test_progress_label_prefers_real_title(self):
-        self.assertEqual(Orchestrator._chapter_progress_label("引言", 0), "引言")
-        self.assertEqual(Orchestrator._chapter_progress_label("第一章", 1), "第一章")
-        self.assertEqual(Orchestrator._chapter_progress_label("", 1), "章节 2")
+        self.assertEqual(TranslationService.chapter_progress_label("引言", 0), "引言")
+        self.assertEqual(TranslationService.chapter_progress_label("第一章", 1), "第一章")
+        self.assertEqual(TranslationService.chapter_progress_label("", 1), "章节 2")
 
     def test_progress_covers_preparation_and_output_stages(self):
         with tempfile.TemporaryDirectory() as d:
