+++ tests/test_review_polish.py
@@ -143,7 +143,7 @@ class TestReviewer(unittest.TestCase):
 
         with tempfile.TemporaryDirectory() as d:
             debug = ReviewRunStore(d)
-            issues = orch._review_chapter(
+            issues = orch._review.review_chapter(
                 [
                     Segment(index=0, source="源文0", target="译文0"),
                     Segment(index=1, source="源文1", target="译文1"),
@@ -229,7 +229,7 @@ class TestReviewer(unittest.TestCase):
 
         with tempfile.TemporaryDirectory() as d:
             debug = ReviewRunStore(d)
-            issues = orch._review_chapter(
+            issues = orch._review.review_chapter(
                 segments,
                 [],
                 chapter_index=7,
@@ -260,7 +260,7 @@ class TestReviewer(unittest.TestCase):
         cfg.pipeline.review_output_retries = 2
         client = FakeClient(handler=handler)
 
-        issues = Orchestrator(cfg, client=client)._review_chapter(
+        issues = Orchestrator(cfg, client=client)._review.review_chapter(
             [Segment(index=0, source="源文", target="译文")],
             [],
         )
@@ -274,7 +274,7 @@ class TestReviewer(unittest.TestCase):
         client = FakeClient(handler=lambda m, t, j: "")
 
         with self.assertRaisesRegex(ReviewOutputError, "malformed_json"):
-            Orchestrator(cfg, client=client)._review_chapter(
+            Orchestrator(cfg, client=client)._review.review_chapter(
                 [Segment(index=0, source="源文", target="译文")],
                 [],
             )
@@ -309,7 +309,7 @@ class TestReviewer(unittest.TestCase):
             Segment(index=1, source="源文乙", target="译文乙"),
         ]
 
-        issues = orch._review_chapter(segments, [])
+        issues = orch._review.review_chapter(segments, [])
 
         self.assertEqual([it["index"] for it in issues], [0, 1])
         self.assertEqual([it["detail"] for it in issues], ["甲", "乙"])
